+++ apps/ymir-sdl3/src/app/events/gui_event.hpp
@@ -79,6 +79,8 @@ struct GUIEvent {
         TryLoadCDBlockROM,
         ReloadCDBlockROM,
 
+        CheckForUpdates,
+
         TakeScreenshot,
 
         // Emulator notifications